feat(app): rooms list + chat route at product quality on phones (#17)
Extract the Rooms tab home (mobile_rooms.dart) and the pushed chat route
(mobile_room.dart) out of mobile_shell.dart, and polish both to the mobile
design bar:

- room rows render the state as dot + label (never color alone) by filling
  the single sidebarRoomMeta message's {stateLabel} slot via templateParts;
  the accent dot + glow stays earned (open session only, sidebar parity)
- 44dp touch floor below the breakpoint, desktop byte-identical (web
  `.btn { min-height: 44px }` parity): create/join affordance rows,
  composer send + attach targets, room-header action buttons
- the chat header caps at ~45% of the route and scrolls internally so
  wrapped French actions / peer chips / a soft-keyboard inset can never
  squeeze the timeline out or hard-overflow the column
- timeline truncation at phone widths: flexible sender names in meta rows,
  the agent-card header wraps its status chip to a second run instead of
  overflowing (spaceBetween keeps the desktop right-flush layout), and the
  file-tile fetch control flexes below the breakpoint
- JeliyaButton labels ellipsize when width-squeezed (web white-space:
  nowrap), never overflow
